print_result print the float collision density

+++ src/devicebridge.cu
@@ -112,11 +112,11 @@ void save_results(unsigned ibat, unsigned gridx, unsigned blockx, unsigned num_s
 }
 
 
-void print_results(unsigned meshes, unsigned nbat, unsigned *tally){
+void print_results(unsigned meshes, unsigned nbat, float *tally){
   int im,ib;
   for(ib=0;ib<nbat;ib++){
     for(im=0;im<meshes;im++){
-      printf("%6d ",tally[ib*meshes+im]);
+      printf("%.5f ",tally[ib*meshes+im]);
     }
     printf("\n");
   }
